pyproject.toml
[project]
name = "peticao-inicial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain==0.3",
    "python-dotenv==1.1",
    "pytest>=8.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "black",
]
